refactor(probe): migrate bgp to per-protocol plugin
